Allows a Column to have unescaped data
This is the first step in allowing custom data like svg or any data that can't be put through htmlspecialchars to appear in the table. It doesn't add any security overhead, all data is going through the {{ }} process, just in the render cell method instead of the blade template. This will now allow you to add chain ->unescaped() onto any declaration.