Add information on creating an action link for a custom page
This took me longer than I care to admit to sort out! Mainly because I was reading both the admin and table builder documents on actions, the table builder uses the Laravel route() helper in actions and this document uses the resource.

I found the answer here: https://github.com/filamentphp/filament/discussions/2282 as I am not the only one to have missed the difference I figured it might be handy in the docs!