Update CanManipulateFiles.php (WINDOWS)
I was creating a theme in Windows and I got errors when creating the files because of the ensureDirectoryExists function. It was checking the last '/' but in WIndows this was a '\' but not always, otherwise you could have used the DIRECTORY_SEPERATOR constant. This fix is better using the pathinfo PHP function