Bug: Global Search Crash when searched in ID
The global search crashes as it is trying to perform lower(id) if global search is implemented for id too. 

Before this fix:
The search crashes with an SQL error:
function lower(bigint) does not exist. This is tested with PostGres.

After this fix:
The search works normally and returns the result normally.