Update 01-installation.md
If a user is trying the installation guide with a new laravel installation, there is a chance that the migration has not been executed yet. But it's necessary for the next step "Create a user".

So, giving him a hint to run migration before trying and failing to create a user.