Add dropdown indicator to topbar items
A dropdown indicator has been added to the topbar items, providing a more intuitive interface for users. This changes the UX to better indicate when a topbar item can be interacted with as a dropdown. A conditional is added to check if the item URL exists, if it doesn't the dropdown indicator is shown.