Change Notifications object naming in Swedish
I don't know if there is an appointed head of Swedish translations. But I would like to propose a change for the naming of "Notifications" in Swedish.

The current "Aviseringar" (in plural form) is a word mor commonly used when issuing formal notifications like sending invoices, or being notified by an authority.

My proposed change "Notiser" is a more direct translation of "Nofitication" and is more common in digital communication and natural language.