only show notifications icon when user has notifications
Previously the notifications icon was always shown, even if there are no notifications: icon with badge (0). The icon should only be shown to indicate action is expected from the user.

With this change the notifications icon is only shown if the user has notifications, either read or unread. 

When pressing "mark all as read" the icon remains, with badge (0). The user can still access old notifications.

When pressing "clear" all notifications disappear, as well as the icon (on the next refresh). That's as expected, as there are no notifications to access.