Crosslink docs for Table Widgets
I find myself repeatedly clicking on the Widgets package documentation looking for Widget documentation for Pages, Resources, Tables, Dashboards ... and then I have to go hunting.
Cross-linking them will hopefully make it less painful for others who are looking for the same.

(Sorry for the renumbering, but it felt really strange having `02-Stats`, `03-Charts`, `04-Adding A Widget To A Blade View`, and then `05-Tables` ... so I renumbered 04 to 05.)