Fixing BelongsToThrough Error
After integrating \Znck\Eloquent\Relations\BelongsToThrough to support this relation when saving a form, an error arises if this relationship is utilized in the form. The error specifically pertains to a 'Call to undefined method Znck\Eloquent\Relations\BelongsToThrough::associate().' Some individuals in the Discord community have also reported encountering this problem. In my opinion, the best suggestion would be to examine the saveRelationshipsUsing() method. It seems that there is no method to prevent this incorrect association, as this relation does not require an 'associate' or 'sync' method, similar to belongsToMany. To resolve this issue, I added some lines in select.php, and since then, the problem has not recurred.