chore: improve type support for model-related methods
Enhanced type support by adding docblock annotations with `class-string<Model>` and, in some cases, adding the native type `Model & Authenticatable` to ensure the object extends the `Model` class and implements the `Authenticatable` interface.

Addressed an edge case for `SpatieTagInput` by checking if the `getTagClassName` method exists before calling it. This was necessary because there isn't a docblock annotation to describe an instance of a class using a specific trait.