Fix: TextColumn size property not applying when badge() is used
TextColumn's `size()` property was not taking effect when `badge()` was applied.
This happened because `badge()` and `size()` are handled as different components.

Now, the `size()` property will correctly apply, even when `badge()` is used.
the extras TextColumnSize::ExtraSmall is set as default as it is right now with one size.