Fix/improve key-value component functionality
Summary of proposed changes
updateRows: I’m using the stateRows array to preserve any changes that the user is making. It’s a copy of the state in array form. The extraRows are rows that the user has entered that are duplicates or empty and therefore not updated to the state yet. Rows are only updated when they don't match the state.
updateState: Sync the rows but only the ones that don’t have an issue. (@finoghentov solution)
Validation: However if a user enters a duplicate key/value, the responsibility of checking this falls on the backend validation. This will not generate an error if a user enters a duplicate key and saves the form, it will silently skip the attempt. If we want front end validation I propose a new method to handle this.