Prevent setting the initial height on every resize
This causes complete relayout + scroll anchoring issues in Firefox when the Textarea is in a flex container and another flex container is above it with scrollable content. The scroll anchoring will break and it will scroll up by the same amount the textarea grew whenever you are at the bottom.

Signed-off-by: Mark Hartman <72379653+Saracaen@users.noreply.github.com>